Abstracting the history manager onto the Memory API
